Check for duplicates server side, not client side.
