+++ client/map_client.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/google/trillian"
 	"github.com/google/trillian/types"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
@@ -56,9 +55,6 @@ func (c *MapClient) GetAndVerifyLatestMapRoot(ctx context.Context) (*types.MapRo
 // GetAndVerifyMapLeaves verifies and returns the requested map leaves.
 // indexes may not contain duplicates.
 func (c *MapClient) GetAndVerifyMapLeaves(ctx context.Context, indexes [][]byte) ([]*trillian.MapLeaf, error) {
-	if err := hasDuplicates(indexes); err != nil {
-		return nil, err
-	}
 	getResp, err := c.Conn.GetLeaves(ctx, &trillian.GetMapLeavesRequest{
 		MapId: c.MapID,
 		Index: indexes,
@@ -73,9 +69,6 @@ func (c *MapClient) GetAndVerifyMapLeaves(ctx context.Context, indexes [][]byte)
 // GetAndVerifyMapLeavesByRevision verifies and returns the requested map leaves at a specific revision.
 // indexes may not contain duplicates.
 func (c *MapClient) GetAndVerifyMapLeavesByRevision(ctx context.Context, revision int64, indexes [][]byte) ([]*trillian.MapLeaf, error) {
-	if err := hasDuplicates(indexes); err != nil {
-		return nil, err
-	}
 	getResp, err := c.Conn.GetLeavesByRevision(ctx, &trillian.GetMapLeavesByRevisionRequest{
 		MapId:    c.MapID,
 		Index:    indexes,
@@ -86,19 +79,6 @@ func (c *MapClient) GetAndVerifyMapLeavesByRevision(ctx context.Context, revisio
 		return nil, status.Errorf(s.Code(), "map.GetLeaves(): %v", s.Message())
 	}
 	return c.VerifyMapLeavesResponse(indexes, revision, getResp)
-}
-
-// hasDuplicates returns an error if there are duplicates in indexes.
-func hasDuplicates(indexes [][]byte) error {
-	set := make(map[string]bool)
-	for _, i := range indexes {
-		if set[string(i)] {
-			return status.Errorf(codes.InvalidArgument,
-				"map.GetLeaves(): index %x requested more than once", i)
-		}
-		set[string(i)] = true
-	}
-	return nil
 }
 
 // SetAndVerifyMapLeaves calls SetLeaves and verifies the signature of the returned map root.
